+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Reduced costs when using address query.
+- Handle and log Google errors when inner response status code differs from OK.
+
+### Added
+- Handle country argument on address suggestions query.
 
 ## [0.3.2] - 2020-12-21
 ### Fixed